+++ src/containerapp/azext_containerapp/commands.py
@@ -277,3 +277,14 @@ def load_command_table(self, args):
         g.custom_command('update', 'update_maintenance_config')
         g.custom_command('remove', 'remove_maintenance_config', confirmation=True)
         g.custom_show_command('list', 'list_maintenance_config')
+
+    with self.command_group('containerapp label-history', is_preview=True) as g:
+        g.custom_show_command('show', 'show_label_history')
+        g.custom_command('list', 'list_label_history')
+
+    with self.command_group('containerapp revision') as g:
+        g.custom_command('set-mode', 'set_revision_mode', exception_handler=ex_handler_factory())
+
+    with self.command_group('containerapp revision label') as g:
+        g.custom_command('add', 'add_revision_label')
+        g.custom_command('remove', 'remove_revision_label')